add loading state to rewards section cards
